Reworked the cleanup process to allow a dangling filelist, such that this dangling entry can be used as the basis for a synthetic filelist. Also added an option to disable synthetic filelists. This fixes #2038. This fixes #2101

+++ Duplicati/Library/Main/Operation/FilelistProcessor.cs
@@ -75,9 +75,10 @@ namespace Duplicati.Library.Main.Operation
         /// <param name="options">The options used</param>
         /// <param name="database">The database to compare with</param>
         /// <param name="log">The log instance to use</param>
-        public static void VerifyRemoteList(BackendManager backend, Options options, LocalDatabase database, IBackendWriter log)
+        /// <param name="protectedfile">A filename that should be excempted for deletion</param>
+        public static void VerifyRemoteList(BackendManager backend, Options options, LocalDatabase database, IBackendWriter log, string protectedfile = null)
         {
-            var tp = RemoteListAnalysis(backend, options, database, log);
+            var tp = RemoteListAnalysis(backend, options, database, log, protectedfile);
             long extraCount = 0;
             long missingCount = 0;
             
@@ -189,10 +190,12 @@ namespace Duplicati.Library.Main.Operation
         /// <param name="backend">The backend instance to use</param>
         /// <param name="options">The options used</param>
         /// <param name="database">The database to compare with</param>
-        public static RemoteAnalysisResult RemoteListAnalysis(BackendManager backend, Options options, LocalDatabase database, IBackendWriter log)
+        /// <param name="protectedfile">A filename that should be excempted for deletion</param>
+        public static RemoteAnalysisResult RemoteListAnalysis(BackendManager backend, Options options, LocalDatabase database, IBackendWriter log, string protectedfile)
         {
             var rawlist = backend.List();
             var lookup = new Dictionary<string, Volumes.IParsedVolume>();
+            protectedfile = protectedfile ?? string.Empty;
 
             var remotelist = (from n in rawlist
                                        let p = Volumes.VolumeBase.ParseFilename(n)
@@ -277,8 +280,15 @@ namespace Duplicati.Library.Main.Operation
                             }
                             else
                             {
-                                log.AddMessage(string.Format("removing file listed as {0}: {1}", i.State, i.Name));
-                                cleanupRemovedRemoteVolumes.Add(i.Name);
+                                if (string.Equals(i.Name, protectedfile) && i.State == RemoteVolumeState.Temporary)
+                                {
+                                    log.AddMessage(string.Format("keeping protected incomplete remote file listed as {0}: {1}", i.State, i.Name));
+                                }
+                                else
+                                {
+                                    log.AddMessage(string.Format("removing file listed as {0}: {1}", i.State, i.Name));
+                                    cleanupRemovedRemoteVolumes.Add(i.Name);
+                                }
                             }
                         }
                         break;
@@ -290,14 +300,30 @@ namespace Duplicati.Library.Main.Operation
                         }
                         else if (!remoteFound)
                         {
-                            log.AddMessage(string.Format("scheduling missing file for deletion, currently listed as {0}: {1}", i.State, i.Name));
-                            cleanupRemovedRemoteVolumes.Add(i.Name);
-                            database.UpdateRemoteVolume(i.Name, RemoteVolumeState.Deleting, i.Size, i.Hash, false, TimeSpan.FromHours(2), null);
+
+                            if (string.Equals(i.Name, protectedfile))
+                            {
+                                log.AddMessage(string.Format("keeping protected incomplete remote file listed as {0}: {1}", i.State, i.Name));
+                                database.UpdateRemoteVolume(i.Name, RemoteVolumeState.Temporary, i.Size, i.Hash, false, new TimeSpan(0), null);
+                            }
+                            else
+                            {
+                                log.AddMessage(string.Format("scheduling missing file for deletion, currently listed as {0}: {1}", i.State, i.Name));
+                                cleanupRemovedRemoteVolumes.Add(i.Name);
+                                database.UpdateRemoteVolume(i.Name, RemoteVolumeState.Deleting, i.Size, i.Hash, false, TimeSpan.FromHours(2), null);
+                            }
                         }
                         else
                         {
-                            log.AddMessage(string.Format("removing incomplete remote file listed as {0}: {1}", i.State, i.Name));
-                            backend.Delete(i.Name, i.Size, true);
+                            if (string.Equals(i.Name, protectedfile))
+                            {
+                                log.AddMessage(string.Format("keeping protected incomplete remote file listed as {0}: {1}", i.State, i.Name));
+                            }
+                            else
+                            {
+                                log.AddMessage(string.Format("removing incomplete remote file listed as {0}: {1}", i.State, i.Name));
+                                backend.Delete(i.Name, i.Size, true);
+                            }
                         }
                         break;
